Added work of the first day to course repo
Before, I had the work in the 100-day-of-code log repo. That did not
make much sense. I might work on other things during the #100DaysOfCode
than this course. So I will log other references. That would get
confusing when there is work in the log repo, but not all of it.

Also, it makes more sense to have the work I do for the course in the
course repo. Otherwise I would just have the the course material on my
disk for not real reason, other than reference. Now, I am actively
making use of the course repo while I am working on the course. That
sounds a lot better.