Added pattern route to pull any *.md from GitHub
To not be required to recreate a new route for every markdown file in
the GitHub log repo, I added one route for any `.md` file. The file
with the same name is then pulled as raw Markdown from my GitHub log
repo.

The GitHub log repo works with relative Markdown links. With this new
route, any Markdown file in the top level of my log repo can be reached.