Prevent creation right for anonymous users
For some reason, adding the `IsOwnerOrReadOnly` permission
to the `ListCreate` view allowed anonymous users to create new
quotes.

Not sure how this happens. My only explanation is that maybe both
`obj.user` and `request.user` default to `None`, which then is equal.
But it does not really make any sense.