Quote user based on request during create
It was possible for users to define another user as the owner of
a quote in in the API front-end. This does not really make sense.
I had to override the creation method of the view to prevent this
and base the quotes user on the request user.

This does not solve the problem. The same action is still possible in
the update view.