Serialize user as hidden field with default
Before, it was possible for logged-in users to set the value for the
owning user of a quote. This means the quote could be assigned to a
different user than the one logged in. This was possible during create
and update operations.

I had overridden the `create` function of the `ListCreateView`, so that
the user input was ignored and the logged-in user would be saved in the
user field.

This was not a very clean solution and would require similar code for
the update view.

There is a more convenient way of achieving this. I was pointed to the
solution on Stack Overflow by Bob Belderbos (teacher of this lesson).

https://stackoverflow.com/a/53193276

Al that is necessary is to define the user filed in the serializer as a
hidden field with a default value.
The current user is retrieved from the field context, which contains the
request.

https://www.django-rest-framework.org/api-guide/fields/#default

This now does not show the user in the listing anymore and automatically
sets is during creation or update to the logged in user.

Maybe there is a way to still show the user, but not allow modification
of the field during create/update.