Extract subheading detection
The previous implementation just looked for the next "Link(s)" header.
This would find a link header even if it is after the next day header.

That behavior has to be prevented. To do so, a custom function was
created that only looks for subheader elements (of the day element)
until it hits the next day header.