Add day offset command line argument
It can happen that the script is run past mid-night, but you still
want to count this as the same day, e.g. because you were working long.

To allow for the script to work in that case, a command line option is
added. This option can be used to create an offset in days to the
current day for which the log should be tweeted.