Move business logic into new main function
Everything that was previously just executed after the if-name-main
construct of the module is now moved into a `main()` function.

This is in preparation of packaging the script and making it
installable. In that step I will need a clear entrypoint into the
script. That is going to be the main function.