Use class to abstract DB interaction
Before, the operational function to create and retrieve URLs needed the
table to perform the operation on explicitly.

Now there is one class that can be instantiated. This class makes sure
that the operations have a table on which to operate. This should
make the handling in the views easier, because I do not need to keep
track of the `table` instance. The functions are only available as
methods of an object that instantiates the table too.