Add host url to short
Before, only the short hand was returned. This is not optimal from a
consumer (of the API) perspective. The consumer would need to keep
track of the host used to create the short link and build the short
link from the short hand and the host url.

Now, the host url is automatically added to the response. This makes
the returned shortlink usable without any additional work on the
consumer side.