Fix create endpoint to actually expect json
Before, I used the same function as for the form processing endpoint.
The processing function tried to access data from the form in both
cases though.

This does not work, when the request is actually json.

The processing function has been fixed to expect the long url to be
passed in directly. The API view extracts it from JSON, while the
form processing endpoint extracted it from posted form data.