Stripping white space from long url
When the long URL is create with trailing white space, this ends up in
the database. When trying to redirect to this URL, it fails because
the headers do not allow for whitespaces or, more specific, newlines
to be contained in the location header for redirects.

This is fixed now by stipping the leading and trailing white spaces
before saving and redirecting in the app.

It probably makes sense to implement this directly in the database
layer too.