Binding host ports to localhost
Before, the port publishing was not bound. This lead to the ports
on the host being opened to the outside world. This was not intended and
may be dangerous!

The binding to the localhost should always be performed to prevent
outside access to the ports through which the host communicates with the
container.

To enable the outside access in production, you can use a firewall that
limits the opened ports and then use a reverse proxy to map from a
specifically opened port in the firewall to the localhost and port that
exposes the container.